backend/pyproject.toml
[project]
name = "jobboard-backend"
version = "0.1.0"
description = "Internal book-cover marketplace and freelancer database"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20",
    "asyncpg>=0.29",
    "boto3>=1.34",
    "python-multipart>=0.0.9",
    "pyjwt[crypto]>=2.8",
    "httpx>=0.27",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py310"
